requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dataclasses==0.6
defusedxml==0.8.0rc2
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
djoser==2.2.3
dnspython==2.6.1
GDAL==3.8.5
geopandas==1.0.1
idna==3.7
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
shapely==2.0.5
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
spatialite==0.0.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
